backend/pyproject.toml
[project]
name = "todo-api"
version = "0.1.0"
description = "DAG-based todo API with SSE subscriptions"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "neo4j>=5.27",
    "python-dotenv>=1.0",
    "pydantic>=2.10",
    "sse-starlette>=2.2",
    "dateparser>=1.2",
    "parsedatetime>=2.6",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28",
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
